+++ README.md
@@ -88,3 +88,5 @@
     - [x] 테이블
     - [x] 단체지정
 - [x] 불필요한 setter를 제거한다.
+- [x] 테스트 픽스쳐 패키지를 분리한다.
+- [ ] 레포지토리를 리팩토링한다.
