+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -1,11 +1,11 @@
 package kitchenpos.application;
 
-import static kitchenpos.application.fixture.MenuGroupFixture.메뉴그룹A;
-import static kitchenpos.application.fixture.MenuGroupFixture.메뉴그룹B;
-import static kitchenpos.application.fixture.ProductFixture.짜장면;
-import static kitchenpos.application.fixture.ProductFixture.짬뽕;
-import static kitchenpos.application.fixture.ProductFixture.탕수육;
-import static kitchenpos.application.fixture.dto.MenuDtoFixture.createMenuRequest;
+import static kitchenpos.fixture.domain.MenuGroupFixture.메뉴그룹A;
+import static kitchenpos.fixture.domain.MenuGroupFixture.메뉴그룹B;
+import static kitchenpos.fixture.domain.ProductFixture.짜장면;
+import static kitchenpos.fixture.domain.ProductFixture.짬뽕;
+import static kitchenpos.fixture.domain.ProductFixture.탕수육;
+import static kitchenpos.fixture.dto.MenuDtoFixture.createMenuRequest;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
