+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -1,15 +1,15 @@
 package kitchenpos.application;
 
-import static kitchenpos.application.fixture.MenuFixture.createMenu;
-import static kitchenpos.application.fixture.MenuGroupFixture.메뉴그룹A;
-import static kitchenpos.application.fixture.MenuGroupFixture.메뉴그룹B;
-import static kitchenpos.application.fixture.OrderTableFixture.createOrderTable;
-import static kitchenpos.application.fixture.ProductFixture.짜장면;
-import static kitchenpos.application.fixture.ProductFixture.탕수육;
-import static kitchenpos.application.fixture.dto.MenuDtoFixture.createMenuRequest;
-import static kitchenpos.application.fixture.dto.OrderDtoFixture.createOrderRequest;
-import static kitchenpos.application.fixture.dto.OrderDtoFixture.forUpdateStatus;
-import static kitchenpos.application.fixture.dto.OrderTableDtoFixture.forUpdateEmpty;
+import static kitchenpos.fixture.domain.MenuFixture.createMenu;
+import static kitchenpos.fixture.domain.MenuGroupFixture.메뉴그룹A;
+import static kitchenpos.fixture.domain.MenuGroupFixture.메뉴그룹B;
+import static kitchenpos.fixture.domain.OrderTableFixture.createOrderTable;
+import static kitchenpos.fixture.domain.ProductFixture.짜장면;
+import static kitchenpos.fixture.domain.ProductFixture.탕수육;
+import static kitchenpos.fixture.dto.MenuDtoFixture.createMenuRequest;
+import static kitchenpos.fixture.dto.OrderDtoFixture.createOrderRequest;
+import static kitchenpos.fixture.dto.OrderDtoFixture.forUpdateStatus;
+import static kitchenpos.fixture.dto.OrderTableDtoFixture.forUpdateEmpty;
 import static kitchenpos.domain.OrderStatus.COOKING;
 import static kitchenpos.domain.OrderStatus.MEAL;
 import static org.assertj.core.api.Assertions.assertThat;
