+++ src/test/java/kitchenpos/fixture/domain/OrderFixture.java
@@ -1,4 +1,4 @@
-package kitchenpos.application.fixture;
+package kitchenpos.fixture.domain;
 
 import java.util.ArrayList;
 import java.util.List;
